% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_go.R
\name{bw_go}
\alias{bw_go}
\title{Perform Gene ontology}
\usage{
bw_go(dp_res, p_val = 0.01, fdr = NULL, diff_class = "all",
  bn = NULL, plotResults = TRUE)
}
\arguments{
\item{dp_res}{results generated from \code{\link{diffpeak}}}

\item{p_val}{Default 0.01}

\item{fdr}{Default NULL}

\item{diff_class}{Default 'all', can be 'up' or 'down'}

\item{bn}{basename for output tsv file}

\item{plotResults}{Default TRUE}
}
\description{
This function performs gene ontology for output generatd from  \code{\link{diffpeak}}. For now only works with 'hg19' or 'mm10'
}
