% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{bw_pca}
\alias{bw_pca}
\title{Groups samples based on signal. Useful for QC analysis}
\usage{
bw_pca(bw_summary = NULL, condition = NULL, top_fract = 0.25,
  pcs = c(1, 2))
}
\arguments{
\item{bw_summary}{output from \code{\link{extract_summary}}}

\item{condition}{a vector of condition for each bigWig file. Samples belonging to same condition are treated as replicates. Default NULL.}

\item{pcs}{Principal components to draw.}

\item{fract}{Fraction of data used for PCA or clustering. Default 0.25}
}
\description{
Groups samples based on signal. Useful for QC analysis
}
