% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bwtool.R
\name{check_bwtools}
\alias{check_bwtools}
\title{Checks for bwtool installation binary under system path or user defined path}
\usage{
check_bwtools(path = NULL, warn = TRUE)
}
\arguments{
\item{path}{path to bwtool binary. Default looks under system directory.}

\item{warn}{whether to warn user with log messages. Default TRUE}
}
\description{
Checks for bwtool installation binary under system path or user defined path
}
\details{
By default looks for bwtool under system path.
}
\examples{
check_bwtools()
}
