% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_profile_plot.R
\name{clustered_profile_plot}
\alias{clustered_profile_plot}
\title{Draw a profile plot}
\usage{
clustered_profile_plot(mat_list, summarizeBy = "mean", color = NULL,
  n_rows = NULL, n_cols = NULL, sample_names = NULL,
  auto_y_axis = TRUE, ci = FALSE, line_size = 2, axis_lwd = 3,
  axis_font_size = 1, legend_font_size = 1)
}
\arguments{
\item{mat_list}{Input matrix list generated by \code{\link{extract_matrix}} run with k}

\item{color}{named vector of colors for each sample}

\item{sample_names}{names for samples. Default NULL parses from bigwig files.}

\item{auto_y_axis}{Default TRUE. If FALSE plots from zero to max.}

\item{ci}{add SEM bars.}

\item{line_size}{line width. Default 3.}

\item{axis_lwd}{axis line width. Default 3.}

\item{axis_font_size}{font size for axis. Default 1.}

\item{legend_font_size}{cex for font size. Default 1.}
}
\description{
Draw a profile plot
}
\details{
This function takes output from clustered extract_matrix and draws a profile plot
}
