% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{diffpeak}
\alias{diffpeak}
\title{Differentail Peak Analysis}
\usage{
diffpeak(bw_summary = NULL, condition = NULL, op_dir = "./",
  num = NULL, den = NULL)
}
\arguments{
\item{bw_summary}{Output from \code{\link{extract_summary}}}

\item{condition}{a column name in \code{coldata} containing sample conditions. Default NULL.}

\item{num}{Numerator condition. Default NULL}

\item{den}{Denominator condition. Default NULL}
}
\description{
Differentail Peak Analysis
}
\details{
Takes output from \code{extract_summary} and performs differential peak analysis with Limma
}
