% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Pol2PI.R
\name{estimate_Pol2PI}
\alias{estimate_Pol2PI}
\title{Estimates polymerase index from Pol2 ChIP seq data.}
\usage{
estimate_Pol2PI(coldata, tx_model = NULL, tssr_up = 50,
  tssr_down = 300, gb = 3000, op_dir = "./")
}
\arguments{
\item{coldata}{coldata Coldata generated from \code{\link{read_coldata}}}

\item{tx_model}{Genes in BED12 format.}

\item{tssr_up}{Region upstream of TSS. Defaulr 300}

\item{gb}{gene body. Region after Transcription end site. Deafult 3000.}

\item{op_dir}{Directory to store results. Defult "./"}

\item{tssr_up}{Region downstream of TSS. Defaulr 50}
}
\description{
Estimates polymerase index from Pol2 ChIP seq data.
}
