% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matrix.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Extract matrix of signal values from given bigWig files around region of interset - bed file.}
\usage{
extract_matrix(coldata = NULL, bed = NULL, genome = NULL,
  binSize = 25, startFrom = "center", up = 2500, down = 2500,
  k = NULL, op_dir = "./", rmAfter = TRUE, bwt_path = NULL,
  nthreads = 4)
}
\arguments{
\item{coldata}{Input coldata generated from \code{\link{read_coldata}}}

\item{bed}{Input bed file}

\item{genome}{Can be hg19 or mm10. Default NULL. If \code{bed} and \code{genome} arguments are given, \code{bed} will be used.}

\item{binSize}{Default 25bp}

\item{startFrom}{region of interest. Default from peak 'center'. Can be 'tss' or 'tes'.}

\item{up}{upsteram in bp. Default 2500.}

\item{down}{downstream in bp. Default 2500.}

\item{k}{cluster regions with k-means where k is the number of clusters. Defualt NULL.}

\item{op_dir}{Directory to store results. Defult "./"}

\item{rmAfter}{remove matrix files generated by bwtool once data is read into R. Default TRUE.}

\item{nthreads}{threads to use. Default 4.}

\item{bw_path}{path to bwtool. Default looks under system path.}
}
\description{
Extract matrix of signal values from given bigWig files around region of interset - bed file.
}
\details{
This function is a wrapper around bwtool matrix command and extracts signals from bigWig files from an input BED file.
Once extraction is complete matrix can be summarized by mean or median signals.
}
