% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary.R
\name{extract_summary}
\alias{extract_summary}
\title{Extract area under the curve for every peak from from given bigWig files.}
\usage{
extract_summary(coldata = NULL, bed = NULL, genome = NULL,
  startFrom = NULL, up = 2500, down = 2500, op_dir = "./",
  rmAfter = TRUE, keepBed = TRUE, bwt_path = NULL,
  bedHeader = FALSE, nthreads = 4, remove_dups = TRUE)
}
\arguments{
\item{coldata}{Coldata generated from \code{\link{read_coldata}}}

\item{bed}{Input bed file. Can also be a gz compressed narrowPeak/BED}

\item{genome}{can be hg19 or mm10}

\item{startFrom}{NULL Default is to estimate complete AUC for each BED entry. Can be 'tss', 'tes' or 'center'}

\item{up}{Default 2500. Only applicable if \code{startFrom} is provided.}

\item{down}{Default 2500. Only applicable if \code{startFrom} is provided.}

\item{op_dir}{Directory to store results. Defult "./"}

\item{rmAfter}{remove matrix files generated by bwtool once data is read into R. Default TRUE.}

\item{keepBed}{Keep output BED as is? Default TRUE}

\item{bedHeader}{Does input BED file has header. Default FALSE}

\item{nthreads}{Threads to use. Default 4.}

\item{remove_dups}{Remove duplicated BED entries. Default TRUE, removes duplicated BED entries with same start and end coordinates.}

\item{bw_path}{path to bwtool. Default looks under system path.}
}
\description{
Extract area under the curve for every peak from from given bigWig files.
}
\details{
This function is a wrapper around bwtool summary command and extracts peak intesities from bigWig files for every peak in an input BED file.
Once extraction is complete signal values are combined into a single data.table
}
