% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_size_dist.R
\name{peak_size}
\alias{peak_size}
\title{Plots peak size distribution}
\usage{
peak_size(bed_files, sample_names = NULL, orderByMedian = TRUE,
  fs = 0.8, show = NULL)
}
\arguments{
\item{bed_files}{path to bed files}

\item{sample_names}{Default NULL. Correspoding sample names}

\item{orderByMedian}{Defaukt TRUE}

\item{fs}{font size}

\item{show}{Default NULL. Can be mean, ,median or N}
}
\description{
Plots peak size distribution
}
