% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Draw heatmap}
\usage{
plot_heatmap(mat_list, sortBy = "mean", rcb_pal = "Blues",
  sample_names = NULL, title_size = 0.8, top_profile = FALSE,
  zmins = NULL, zmaxs = NULL, scale = FALSE, file_name = NULL,
  hm_width = NULL, hm_height = 12, ...)
}
\arguments{
\item{mat_list}{Input matrix list generated by \code{\link{extract_matrix}}}

\item{sortBy}{Sort matrix by.., Can be mean, median. Default mean.}

\item{rcb_pal}{Color palettte to use. RColorBrewer::display.brewer.all() to see available palettes}

\item{title_size}{size of title. Default 0.8}

\item{top_profile}{Boolean. Whether to draw top profile plot.}

\item{zmins}{Manually specify min scores to include in heatmap}

\item{zmaxs}{Manually specify max scores to include in heatmap}

\item{scale}{Whether to row scale the matrix. Default FALSE}

\item{file_name}{Default NULL. If provided saves plot as a tiff.}

\item{hm_width}{Width of the plot}

\item{hm_height}{Height of the plot}

\item{sample_bames}{Manually specifcy sample names.}
}
\value{
If matrix was generated with `k`, returns BED file with cluster informartion
}
\description{
Draw heatmap
}
\details{
This function takes output from extract_matrix and draws a heatmap
}
