% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Draw a profile plot}
\usage{
plot_profile(mat_list, summarizeBy = "mean", color = NULL,
  ci = FALSE, legend_font_size = 1, condition = NULL,
  collapse_reps = TRUE, line_size = 2, axis_lwd = 3,
  axis_font_size = 1, sample_names = NULL, auto_y_axis = TRUE,
  y_lims = NULL)
}
\arguments{
\item{mat_list}{Input matrix list generated by \code{\link{extract_matrix}}}

\item{summarizeBy}{summarize matrix by mean or median. Default mean.}

\item{color}{named vector of colors for each sample}

\item{ci}{add SEM bars.}

\item{legend_font_size}{cex for font size. Default 1.}

\item{condition}{a column name in \code{coldata} containing sample conditions. Default NULL.}

\item{collapse_reps}{Samples belonging to same condition to estimate average signal. Default TRUE if \code{condition} is given.}

\item{line_size}{line width. Default 3.}

\item{axis_lwd}{axis line width. Default 3.}

\item{axis_font_size}{font size for axis. Default 1.}

\item{sample_names}{names for samples. Default NULL parses from bigwig files.}

\item{auto_y_axis}{Default TRUE. If FALSE plots from zero to max.}

\item{y_lims}{Default NULL. Manually specify Y-limits.}
}
\description{
Draw a profile plot
}
\details{
This function takes output from extract_matrix and draws a profile plot
}
