% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tracks.R
\name{plot_tracks}
\alias{plot_tracks}
\title{Draw a track view of geiven genomic loci}
\usage{
plot_tracks(bigWigs = NULL, chr, start, end, op_dir = "./",
  rmAfter = TRUE, bwt_path = NULL)
}
\arguments{
\item{mat_list}{Input matrix list generated by \code{\link{extract_matrix}}}

\item{color}{named vector of colors for each sample}
}
\description{
Draw a track view of geiven genomic loci
}
\details{
This function takes output from extract_matrix and draws a profile plot
}
