% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go.R
\name{plotgo}
\alias{plotgo}
\title{This function plots top ten GO process from GOSEq results}
\usage{
plotgo(gores, top = 10, font_col = "#4D004B", bar_col = "#BDD7E7",
  font_size = 0.75)
}
\arguments{
\item{gores}{results generated from \code{\link{bw_go}}}

\item{top}{Default 10}

\item{font_col}{text font color}

\item{bar_col}{color for bars}

\item{font_size}{font size Default 0.75}
}
\description{
This function plots top ten GO process from GOSEq results
}
