% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSE.R
\name{plotSE}
\alias{plotSE}
\title{Draw hocky-stick plot from super enhancer results.}
\usage{
plotSE(enhToGene, markGenes = NULL, keepHighestSE = TRUE)
}
\arguments{
\item{enhToGene}{ENHANCER_TO_GENE.txt file generated by ROSE_GeneMapper.}

\item{markGenes}{SE genes to highlight}

\item{keepHighestSE}{If multiple genes with name are found to be in SE list, default is to keep top ranking entry. Default TRUE.}
}
\description{
Draw hocky-stick plot from super enhancer results.
}
