% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coldata.R
\name{read_coldata}
\alias{read_coldata}
\title{Prepares meta data table from bigWig files.
Output from this fucntion is passed to all downstream functions.}
\usage{
read_coldata(coldata = NULL, files_idx = NULL, sample_idx = NULL)
}
\arguments{
\item{coldata}{a data.frame or a tsv file contaning absolute path to bigWig files and other optional associated sample annotations}

\item{files_idx}{column index containing path to bigWig files. Required.}

\item{sample_idx}{column index containing sample names for corresponding bigWig files. Optional. Default NULL - creates one from bigWig files.}
}
\description{
Prepares meta data table from bigWig files.
Output from this fucntion is passed to all downstream functions.
}
