% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_homer.R
\name{summarize_homer_annots}
\alias{summarize_homer_annots}
\title{Parse peak annotations generated by homer annotatePeaks.pl}
\usage{
summarize_homer_annots(anno, sample_names = NULL, plot = TRUE,
  legend_font_size = 1, label_size = 0.8)
}
\arguments{
\item{anno}{Raw annotations generated by annotatePeaks.pl.}

\item{plot}{If TRUE generates a pie chart for peak distribution}

\item{legend_font_size}{font size for legend. Default 1.}

\item{label_size}{font size for labels. Default 0.8.}
}
\description{
Parse peak annotations generated by homer annotatePeaks.pl
}
\details{
summarizes peak annotations generated with homer annotatePeaks.pl, also a generates a pie chart of peak distributions.
}
